Sdcore-508: Adding support for data plane in gnbsim

+++ transportcommon/transportpeer.go
@@ -7,5 +7,5 @@ package transportcommon
 
 type TransportPeer interface {
 	GetIpAddr() string
-	GetPort() uint16
+	GetPort() int
 }
